Cleaning console.logs

+++ lib/actions/user.actions.ts
@@ -45,7 +45,7 @@ export default async function signInWithCredentials(
 //Sign out user
 export async function signOutUser() {
   const currentCart = await getMyCart();
-  await prisma.cart.delete({ where: { id: currentCart?.id } });
+  if(currentCart) await prisma.cart.delete({ where: { id: currentCart?.id } });
   await signOut();
 }
